fixed cancelFollow and rejectFollow request, updated API to include cancelFollow
